Split the trade calculation engine into buy, sell, and rebalance paths
BuyAllocation, SellPath, and RebalancePath now own their calculation
clusters as static collaborators in the existing package style, with
netInvestable and the model weight map computed once in the engine and
passed in. Rebalance reuses the buy-side hard-limit redistribution and
the sell-side normalized targets through real cross-class calls instead
of duplication. Converting applyLimits from a stream to a loop restored
a call edge the cohesion metric could not see through the lambda, which
also fixed a pre-existing disconnected component. The engine test is
split per path with every assertion preserved verbatim.

God classes 3 -> 2 (both remaining are accepted floors), PMD violations
131 -> 128, top-15 class lines 6305 -> 6033, disconnected classes
452 -> 451 with the baseline lowered to match.
