+++ metrics/scorecard.json
@@ -6,16 +6,16 @@
   "compilerWarnings": 0,
   "deepNestingViolations": 0,
   "disabledTests": 5,
-  "disconnectedClasses": 452,
+  "disconnectedClasses": 451,
   "fqnViolations": 0,
-  "godClasses": 3,
+  "godClasses": 2,
   "interactionVerifyingTestClasses": 363,
-  "interactionVerifyingTestShare": 36.1,
+  "interactionVerifyingTestShare": 36.0,
   "lineCoverage": 84.04,
-  "longClasses": 1,
+  "longClasses": 0,
   "meanInstability": 0.473,
-  "mocklessTestClasses": 462,
-  "mocklessTestShare": 46.0,
+  "mocklessTestClasses": 465,
+  "mocklessTestShare": 46.1,
   "moduleCycles": 0,
   "moduleSliceTests": 0,
   "modulithViolations": 0,
@@ -26,23 +26,23 @@
   "overMockedTestClasses": 113,
   "overMockedTestShare": 11.2,
   "pitestSlicedModules": 52,
-  "pmdViolations": 131,
+  "pmdViolations": 128,
   "pmdViolationsByRule": {
     "CognitiveComplexity": 20,
     "CouplingBetweenObjects": 34,
-    "CyclomaticComplexity": 70,
-    "GodClass": 3,
-    "TooManyMethods": 4
+    "CyclomaticComplexity": 69,
+    "GodClass": 2,
+    "TooManyMethods": 3
   },
-  "slowest10TestClassesSeconds": 117.2,
+  "slowest10TestClassesSeconds": 119.4,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 201,
-  "testClassesOver10Seconds": 5,
-  "testClassesTotal": 1005,
-  "testIntegrationShare": 6.7,
+  "suiteWallTimeSeconds": 207,
+  "testClassesOver10Seconds": 4,
+  "testClassesTotal": 1008,
+  "testIntegrationShare": 6.6,
   "testSliceShare": 13.7,
-  "testUnitShare": 79.6,
-  "top15ClassLines": 6305,
+  "testUnitShare": 79.7,
+  "top15ClassLines": 6033,
   "totalPackages": 227,
   "unmarkedPackages": 0
 }
