+++ src/test/groovy/ee/tuleva/onboarding/investment/transaction/calculation/TradeCalculationEngineTest.java
@@ -1,8 +1,6 @@
 package ee.tuleva.onboarding.investment.transaction.calculation;
 
 import static ee.tuleva.onboarding.fund.TulevaFund.TUV100;
-import static ee.tuleva.onboarding.investment.transaction.CalculationWarningType.REBALANCE_NET_CASH_MISMATCH;
-import static ee.tuleva.onboarding.investment.transaction.CalculationWarningType.REBALANCE_NET_NOT_ACHIEVED;
 import static ee.tuleva.onboarding.investment.transaction.LimitStatus.*;
 import static ee.tuleva.onboarding.investment.transaction.TransactionMode.*;
 import static java.math.BigDecimal.ZERO;
@@ -43,76 +41,6 @@ class TradeCalculationEngineTest {
   }
 
   @Test
-  void buy_withFreeCashBelowThreshold_recordsFreeCashNoTradeReason() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("500000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("1000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    assertThat(result.trades())
-        .allSatisfy(trade -> assertThat(trade.tradeAmount()).isEqualByComparingTo(ZERO));
-    assertThat(result.noTradeReason())
-        .isEqualTo(
-            "No trades: mode=BUY, reason=freeCashBelowMinTransactionThreshold, freeCash=1000,"
-                + " minTransactionThreshold=5000");
-  }
-
-  @Test
-  void sell_withoutCashShortfall_recordsNoCashShortfallReason() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("500000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.noTradeReason())
-        .isEqualTo("No trades: mode=SELL, reason=noCashShortfall, freeCash=0");
-  }
-
-  @Test
-  void rebalance_withNoDrift_recordsNoDriftReason() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("1000000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    assertThat(result.noTradeReason())
-        .isEqualTo("No trades: mode=REBALANCE, reason=noDriftBeyondThreshold");
-  }
-
-  @Test
   void calculate_withTrades_hasNoNoTradeReason() {
     var input =
         FundTransactionInput.builder()
@@ -131,1748 +59,6 @@ class TradeCalculationEngineTest {
     var result = engine.calculate(input, BUY);
 
     assertThat(result.noTradeReason()).isNull();
-  }
-
-  @Test
-  void buy_allocatesFreeCashToUnderweightPositions() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("300000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("200000")),
-                    new PositionSnapshot("IE00C", new BigDecimal("100000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.40")),
-                    new ModelWeight("IE00B", new BigDecimal("0.35")),
-                    new ModelWeight("IE00C", new BigDecimal("0.25"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                        new PositionLimitSnapshot(new BigDecimal("0.50"), new BigDecimal("0.60")),
-                    "IE00B",
-                        new PositionLimitSnapshot(new BigDecimal("0.50"), new BigDecimal("0.60")),
-                    "IE00C",
-                        new PositionLimitSnapshot(new BigDecimal("0.50"), new BigDecimal("0.60"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    assertThat(result.fund()).isEqualTo(TUV100);
-    assertThat(result.mode()).isEqualTo(BUY);
-    assertThat(result.trades()).hasSize(3);
-
-    BigDecimal totalTradeAmount =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalTradeAmount)
-        .isCloseTo(
-            new BigDecimal("100000"), org.assertj.core.data.Offset.offset(new BigDecimal("1")));
-
-    result
-        .trades()
-        .forEach(
-            trade -> {
-              assertThat(trade.tradeAmount()).isGreaterThanOrEqualTo(ZERO);
-              assertThat(trade.limitStatus()).isEqualTo(OK);
-            });
-  }
-
-  @Test
-  void buy_withZeroFreeCash_returnsZeroTrades() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("500000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    assertThat(result.trades()).hasSize(1);
-    assertThat(result.trades().getFirst().tradeAmount()).isEqualByComparingTo(ZERO);
-  }
-
-  @Test
-  void buy_removesPositionsBelowThreshold() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("400000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("549000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("10000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var nonZeroTrades =
-        result.trades().stream().filter(t -> t.tradeAmount().compareTo(ZERO) > 0).toList();
-    nonZeroTrades.forEach(
-        trade -> assertThat(trade.tradeAmount()).isGreaterThanOrEqualTo(new BigDecimal("5000")));
-  }
-
-  @Test
-  void buy_clipsAtHardLimit() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("580000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("20000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.80")),
-                    new ModelWeight("IE00B", new BigDecimal("0.20"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("200000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                        new PositionLimitSnapshot(new BigDecimal("0.55"), new BigDecimal("0.60")),
-                    "IE00B",
-                        new PositionLimitSnapshot(new BigDecimal("0.50"), new BigDecimal("0.60"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("19900.00"));
-    assertThat(tradeA.projectedWeight()).isLessThanOrEqualTo(new BigDecimal("0.6001"));
-    assertThat(tradeA.limitStatus()).isEqualTo(SOFT_LIMIT_EXCEEDED);
-
-    BigDecimal totalInvested =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalInvested)
-        .isCloseTo(new BigDecimal("200000"), org.assertj.core.data.Offset.offset(BigDecimal.ONE));
-  }
-
-  @Test
-  void buy_excludesInstrumentsWithHeadroomBelowThresholdAndInvestsFullAmount() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("595000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("200000")),
-                    new PositionSnapshot("IE00C", new BigDecimal("100000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.65")),
-                    new ModelWeight("IE00B", new BigDecimal("0.20")),
-                    new ModelWeight("IE00C", new BigDecimal("0.15"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                    new PositionLimitSnapshot(new BigDecimal("0.55"), new BigDecimal("0.60"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(ZERO);
-
-    BigDecimal totalInvested =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalInvested)
-        .isCloseTo(new BigDecimal("100000"), org.assertj.core.data.Offset.offset(BigDecimal.ONE));
-  }
-
-  @Test
-  void buy_redistributesHardLimitExcessToRunnerUpInstruments() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("270000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("300000")),
-                    new PositionSnapshot("IE00C", new BigDecimal("200000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.35")),
-                    new ModelWeight("IE00B", new BigDecimal("0.40")),
-                    new ModelWeight("IE00C", new BigDecimal("0.25"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("180000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                    new PositionLimitSnapshot(new BigDecimal("0.28"), new BigDecimal("0.30"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    assertThat(tradeA.tradeAmount())
-        .isCloseTo(new BigDecimal("29900"), org.assertj.core.data.Offset.offset(BigDecimal.ONE));
-
-    BigDecimal totalInvested =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalInvested)
-        .isCloseTo(new BigDecimal("180000"), org.assertj.core.data.Offset.offset(BigDecimal.ONE));
-  }
-
-  @Test
-  void buy_waterFillsHardLimitExcessAcrossRunnersInsteadOfStrandingCash() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00X", new BigDecimal("240000")),
-                    new PositionSnapshot("IE00P", new BigDecimal("130000")),
-                    new PositionSnapshot("IE00Q", new BigDecimal("130000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00X", new BigDecimal("0.60")),
-                    new ModelWeight("IE00P", new BigDecimal("0.20")),
-                    new ModelWeight("IE00Q", new BigDecimal("0.16"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00X",
-                        new PositionLimitSnapshot(new BigDecimal("0.25"), new BigDecimal("0.26")),
-                    "IE00P",
-                        new PositionLimitSnapshot(new BigDecimal("0.14"), new BigDecimal("0.15")),
-                    "IE00Q",
-                        new PositionLimitSnapshot(new BigDecimal("0.39"), new BigDecimal("0.40"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    BigDecimal totalInvested =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalInvested)
-        .isCloseTo(new BigDecimal("100000"), org.assertj.core.data.Offset.offset(BigDecimal.ONE));
-
-    result
-        .trades()
-        .forEach(
-            trade ->
-                assertThat(trade.projectedWeight())
-                    .isLessThanOrEqualTo(input.positionLimits().get(trade.isin()).hardLimit()));
-  }
-
-  @Test
-  void buy_topsUpBestRunnerWithSubThresholdHardLimitLeftover() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("400000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("400000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("50000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                    new PositionLimitSnapshot(new BigDecimal("0.40"), new BigDecimal("0.4221"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("22000.00"));
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(new BigDecimal("28000.00"));
-
-    BigDecimal totalInvested =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalInvested).isEqualByComparingTo(new BigDecimal("50000.00"));
-  }
-
-  @Test
-  void buy_strandsExcessWhenEveryRunnerHitsItsHardLimit() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("100000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("50000")),
-                    new PositionSnapshot("IE00C", new BigDecimal("50000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.80")),
-                    new ModelWeight("IE00B", new BigDecimal("0.10")),
-                    new ModelWeight("IE00C", new BigDecimal("0.10"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                        new PositionLimitSnapshot(new BigDecimal("0.10"), new BigDecimal("0.13")),
-                    "IE00B",
-                        new PositionLimitSnapshot(new BigDecimal("0.05"), new BigDecimal("0.056")),
-                    "IE00C",
-                        new PositionLimitSnapshot(new BigDecimal("0.05"), new BigDecimal("0.056"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-    var tradeC =
-        result.trades().stream().filter(t -> t.isin().equals("IE00C")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("29900.00"));
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(new BigDecimal("5900.00"));
-    assertThat(tradeC.tradeAmount()).isEqualByComparingTo(new BigDecimal("5900.00"));
-
-    BigDecimal totalInvested =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalInvested).isEqualByComparingTo(new BigDecimal("41700.00"));
-    assertThat(totalInvested).isLessThan(input.freeCash());
-
-    result
-        .trades()
-        .forEach(
-            trade ->
-                assertThat(trade.projectedWeight())
-                    .isLessThanOrEqualTo(input.positionLimits().get(trade.isin()).hardLimit()));
-  }
-
-  @Test
-  void buy_stopsWaterFillOnceLeftoverExcessIsFullyAbsorbed() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("300000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("300000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.60")),
-                    new ModelWeight("IE00B", new BigDecimal("0.40"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("50000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                        new PositionLimitSnapshot(new BigDecimal("0.30"), new BigDecimal("0.31")),
-                    "IE00B",
-                        new PositionLimitSnapshot(
-                            new BigDecimal("0.33"), new BigDecimal("0.3402"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("9900.00"));
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(new BigDecimal("40100.00"));
-
-    BigDecimal totalInvested =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalInvested).isEqualByComparingTo(new BigDecimal("50000.00"));
-  }
-
-  @Test
-  void buy_ignoresOnTargetPositionsWhenRedistributingHardLimitExcess() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("350000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("200000")),
-                    new PositionSnapshot("IE00C", new BigDecimal("300000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.40")),
-                    new ModelWeight("IE00B", new BigDecimal("0.30")),
-                    new ModelWeight("IE00C", new BigDecimal("0.30"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("50000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                    new PositionLimitSnapshot(new BigDecimal("0.35"), new BigDecimal("0.36"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-    var tradeC =
-        result.trades().stream().filter(t -> t.isin().equals("IE00C")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("9900.00"));
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(new BigDecimal("40100.00"));
-    assertThat(tradeC.tradeAmount()).isEqualByComparingTo(ZERO);
-
-    BigDecimal totalInvested =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalInvested).isEqualByComparingTo(new BigDecimal("50000.00"));
-  }
-
-  @Test
-  void buy_flagsSoftLimitExceeded() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("400000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("200000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                    new PositionLimitSnapshot(new BigDecimal("0.50"), new BigDecimal("0.70"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var trade = result.trades().getFirst();
-    assertThat(trade.limitStatus()).isEqualTo(SOFT_LIMIT_EXCEEDED);
-  }
-
-  @Test
-  void sell_sellsOverweightPositions() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("600000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("300000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                        new PositionLimitSnapshot(new BigDecimal("0.55"), new BigDecimal("0.65")),
-                    "IE00B",
-                        new PositionLimitSnapshot(new BigDecimal("0.50"), new BigDecimal("0.60"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.trades()).hasSize(2);
-
-    BigDecimal totalSold =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalSold).isEqualByComparingTo(new BigDecimal("-100000"));
-  }
-
-  @Test
-  void sell_targetsOnlyOverweightPositionsAndLeavesOnTargetPositionUntouched() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("500000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("400000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("900000"))
-            .cashBuffer(new BigDecimal("100000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("-100000"));
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(ZERO);
-  }
-
-  @Test
-  void sell_withPositiveFreeCash_returnsZeroTrades() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("500000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("50000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    result.trades().forEach(trade -> assertThat(trade.tradeAmount()).isEqualByComparingTo(ZERO));
-  }
-
-  @Test
-  void sell_prefersPositionsOverSoftLimit() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("600000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("200000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-50000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                        new PositionLimitSnapshot(new BigDecimal("0.55"), new BigDecimal("0.65")),
-                    "IE00B",
-                        new PositionLimitSnapshot(new BigDecimal("0.50"), new BigDecimal("0.60"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isLessThan(ZERO);
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(ZERO);
-  }
-
-  @Test
-  void sell_capsOverSoftReliefAtModelAndSpillsExcessToOtherOverweightFunds() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("550000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("350000")),
-                    new PositionSnapshot("IE00C", new BigDecimal("100000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.40")),
-                    new ModelWeight("IE00B", new BigDecimal("0.30")),
-                    new ModelWeight("IE00C", new BigDecimal("0.30"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-200000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                        new PositionLimitSnapshot(new BigDecimal("0.50"), new BigDecimal("0.60")),
-                    "IE00B",
-                        new PositionLimitSnapshot(new BigDecimal("0.40"), new BigDecimal("0.50"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.trades())
-        .usingRecursiveComparison()
-        .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
-        .isEqualTo(
-            List.of(
-                new TradeCalculation(
-                    "IE00A", new BigDecimal("-150000"), new BigDecimal("0.400000"), OK),
-                new TradeCalculation(
-                    "IE00B", new BigDecimal("-50000"), new BigDecimal("0.300000"), OK),
-                new TradeCalculation("IE00C", ZERO, new BigDecimal("0.100000"), OK)));
-  }
-
-  @Test
-  void sellFast_sellsFastTaggedInstrumentsFirst() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00FAST", new BigDecimal("200000")),
-                    new PositionSnapshot("IE00SLOW", new BigDecimal("300000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00FAST", new BigDecimal("0.40")),
-                    new ModelWeight("IE00SLOW", new BigDecimal("0.60"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of("IE00FAST"))
-            .build();
-
-    var result = engine.calculate(input, SELL_FAST);
-
-    var fastTrade =
-        result.trades().stream().filter(t -> t.isin().equals("IE00FAST")).findFirst().orElseThrow();
-    assertThat(fastTrade.tradeAmount()).isLessThan(ZERO);
-
-    BigDecimal totalSold =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalSold).isEqualByComparingTo(new BigDecimal("-100000"));
-  }
-
-  @Test
-  void sellFast_overflowsToSlowInstruments() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00FAST", new BigDecimal("50000")),
-                    new PositionSnapshot("IE00SLOW", new BigDecimal("300000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00FAST", new BigDecimal("0.10")),
-                    new ModelWeight("IE00SLOW", new BigDecimal("0.90"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of("IE00FAST"))
-            .build();
-
-    var result = engine.calculate(input, SELL_FAST);
-
-    var fastTrade =
-        result.trades().stream().filter(t -> t.isin().equals("IE00FAST")).findFirst().orElseThrow();
-    assertThat(fastTrade.tradeAmount()).isEqualByComparingTo(new BigDecimal("-50000"));
-
-    var slowTrade =
-        result.trades().stream().filter(t -> t.isin().equals("IE00SLOW")).findFirst().orElseThrow();
-    assertThat(slowTrade.tradeAmount()).isEqualByComparingTo(new BigDecimal("-50000"));
-  }
-
-  @Test
-  void sellFast_sellsMostOverweightInstrumentsFirst() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("500000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("400000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.45")),
-                    new ModelWeight("IE00B", new BigDecimal("0.45"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-25000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of("IE00A", "IE00B"))
-            .build();
-
-    var result = engine.calculate(input, SELL_FAST);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("-25000.00"));
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(ZERO);
-  }
-
-  @Test
-  void sellFast_withPositiveFreeCash_returnsZeroTrades() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("200000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("50000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of("IE00A"))
-            .build();
-
-    var result = engine.calculate(input, SELL_FAST);
-
-    result.trades().forEach(trade -> assertThat(trade.tradeAmount()).isEqualByComparingTo(ZERO));
-  }
-
-  @Test
-  void rebalance_movesPositionsTowardModelWeights() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("600000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("200000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isLessThan(ZERO);
-    assertThat(tradeB.tradeAmount()).isGreaterThan(ZERO);
-  }
-
-  @Test
-  void rebalance_normalizesWeights() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("500000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("400000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.60")),
-                    new ModelWeight("IE00B", new BigDecimal("0.40"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("70000"));
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(new BigDecimal("-20000"));
-  }
-
-  @Test
-  void rebalance_eliminatesSubThresholdTradesAndRedistributes() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("350000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("412000")),
-                    new PositionSnapshot("IE00C", new BigDecimal("188000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.40")),
-                    new ModelWeight("IE00B", new BigDecimal("0.40")),
-                    new ModelWeight("IE00C", new BigDecimal("0.20"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-    var tradeC =
-        result.trades().stream().filter(t -> t.isin().equals("IE00C")).findFirst().orElseThrow();
-
-    assertThat(tradeC.tradeAmount()).isEqualByComparingTo(ZERO);
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("32000.00"));
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(new BigDecimal("-32000.00"));
-  }
-
-  @Test
-  void rebalance_producesNoBuysWhenNoPositionIsUnderTarget() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("600000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("500000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(new BigDecimal("-100000"));
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(ZERO);
-  }
-
-  @Test
-  void rebalance_neverSellsAnyPositionBeyondItsMarketValue() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("100000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("100000"))
-            .cashBuffer(ZERO)
-            .liabilities(new BigDecimal("150000"))
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isGreaterThanOrEqualTo(new BigDecimal("-100000"));
-  }
-
-  @Test
-  void rebalance_neverInflatesSurvivingSellAllocationBeyondMarketValueAfterThresholdElimination() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("3000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("100000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("103000"))
-            .cashBuffer(ZERO)
-            .liabilities(new BigDecimal("203000"))
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeB.tradeAmount()).isGreaterThanOrEqualTo(new BigDecimal("-100000"));
-  }
-
-  @Test
-  void rebalance_waterFillsHardLimitBuyExcessAcrossRunnersInsteadOfStrandingCash() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("100000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("100000")),
-                    new PositionSnapshot("IE00C", new BigDecimal("300000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.30")),
-                    new ModelWeight("IE00C", new BigDecimal("0.20"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                    new PositionLimitSnapshot(new BigDecimal("0.32"), new BigDecimal("0.32"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    BigDecimal totalBought =
-        result.trades().stream()
-            .map(TradeCalculation::tradeAmount)
-            .filter(amount -> amount.compareTo(ZERO) > 0)
-            .reduce(ZERO, BigDecimal::add);
-
-    assertThat(totalBought).isEqualByComparingTo(new BigDecimal("600000"));
-    assertThat(result.trades())
-        .allSatisfy(trade -> assertThat(trade.limitStatus()).isNotEqualTo(HARD_LIMIT_EXCEEDED));
-  }
-
-  @Test
-  void sell_stopsWhenNeedFullyCoveredByExactMarketValueCap() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00OVER", new BigDecimal("40000")),
-                    new PositionSnapshot("IE00ATTARGET", new BigDecimal("100000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00OVER", ZERO),
-                    new ModelWeight("IE00ATTARGET", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-40000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    var overTrade =
-        result.trades().stream().filter(t -> t.isin().equals("IE00OVER")).findFirst().orElseThrow();
-    var atTargetTrade =
-        result.trades().stream()
-            .filter(t -> t.isin().equals("IE00ATTARGET"))
-            .findFirst()
-            .orElseThrow();
-
-    assertThat(overTrade.tradeAmount()).isEqualByComparingTo(new BigDecimal("-40000"));
-    assertThat(atTargetTrade.tradeAmount()).isEqualByComparingTo(ZERO);
-  }
-
-  @Test
-  void sellFast_leavesShortfallWhenNoSlowInstrumentsRemain() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00FAST", new BigDecimal("50000"))))
-            .modelWeights(List.of(new ModelWeight("IE00FAST", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-80000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of("IE00FAST"))
-            .build();
-
-    var result = engine.calculate(input, SELL_FAST);
-
-    var fastTrade =
-        result.trades().stream().filter(t -> t.isin().equals("IE00FAST")).findFirst().orElseThrow();
-    assertThat(fastTrade.tradeAmount()).isEqualByComparingTo(new BigDecimal("-50000"));
-
-    BigDecimal totalSold =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalSold).isEqualByComparingTo(new BigDecimal("-50000"));
-  }
-
-  @Test
-  void emptyPositions_returnsEmptyTrades() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of())
-            .modelWeights(List.of())
-            .grossPortfolioValue(ZERO)
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    assertThat(result.trades()).isEmpty();
-  }
-
-  @Test
-  void singlePosition_getsFullAllocation() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("400000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    assertThat(result.trades()).hasSize(1);
-    assertThat(result.trades().getFirst().tradeAmount())
-        .isEqualByComparingTo(new BigDecimal("100000"));
-  }
-
-  @Test
-  void buy_positionWithNoModelWeight_getsZeroAllocation() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("400000")),
-                    new PositionSnapshot("IE00UNKNOWN", new BigDecimal("100000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var unknownTrade =
-        result.trades().stream()
-            .filter(t -> t.isin().equals("IE00UNKNOWN"))
-            .findFirst()
-            .orElseThrow();
-    assertThat(unknownTrade.tradeAmount()).isEqualByComparingTo(ZERO);
-  }
-
-  @Test
-  void sell_fallbackToStandardWhenAllWithinSoftLimit() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("500000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("400000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-50000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                        new PositionLimitSnapshot(new BigDecimal("0.60"), new BigDecimal("0.70")),
-                    "IE00B",
-                        new PositionLimitSnapshot(new BigDecimal("0.60"), new BigDecimal("0.70"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    BigDecimal totalSold =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalSold).isEqualByComparingTo(new BigDecimal("-50000"));
-
-    var nonZeroTrades =
-        result.trades().stream().filter(t -> t.tradeAmount().compareTo(ZERO) != 0).toList();
-    assertThat(nonZeroTrades).isNotEmpty();
-  }
-
-  @Test
-  void buy_deploysFreeCashWhenAllPositionsAbovePostCashTarget() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("245000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("245000")),
-                    new PositionSnapshot("IE00C", new BigDecimal("245000")),
-                    new PositionSnapshot("IE00D", new BigDecimal("245000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.25")),
-                    new ModelWeight("IE00B", new BigDecimal("0.25")),
-                    new ModelWeight("IE00C", new BigDecimal("0.25")),
-                    new ModelWeight("IE00D", new BigDecimal("0.25"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("50000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("20000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    BigDecimal totalBuy =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalBuy)
-        .isCloseTo(new BigDecimal("20000"), org.assertj.core.data.Offset.offset(BigDecimal.ONE));
-  }
-
-  @Test
-  void buy_fallbackPrefersLeastOverweightFund() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("238000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("250000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("500000"))
-            .cashBuffer(new BigDecimal("25000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("12000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isGreaterThan(tradeB.tradeAmount());
-
-    BigDecimal totalBuy =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalBuy)
-        .isCloseTo(new BigDecimal("12000"), org.assertj.core.data.Offset.offset(BigDecimal.ONE));
-  }
-
-  @Test
-  void buy_includesReceivablesInTargetBaseForScoring() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("90000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("80000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("170000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .receivables(new BigDecimal("60000"))
-            .freeCash(new BigDecimal("20000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-
-    assertThat(tradeA.tradeAmount()).isGreaterThan(ZERO);
-    assertThat(tradeB.tradeAmount()).isGreaterThan(tradeA.tradeAmount());
-
-    BigDecimal totalBuy =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalBuy)
-        .isCloseTo(new BigDecimal("20000"), org.assertj.core.data.Offset.offset(BigDecimal.ONE));
-  }
-
-  @Test
-  void sellFast_redistributesMarketValueCapShortfallWithinFastBucket() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00SMALL", new BigDecimal("10000")),
-                    new PositionSnapshot("IE00BIG", new BigDecimal("100000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00SMALL", ZERO),
-                    new ModelWeight("IE00BIG", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-15000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of("IE00SMALL", "IE00BIG"))
-            .build();
-
-    var result = engine.calculate(input, SELL_FAST);
-
-    var smallTrade =
-        result.trades().stream()
-            .filter(t -> t.isin().equals("IE00SMALL"))
-            .findFirst()
-            .orElseThrow();
-    assertThat(smallTrade.tradeAmount()).isGreaterThanOrEqualTo(new BigDecimal("-10000"));
-
-    BigDecimal totalSold =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalSold).isEqualByComparingTo(new BigDecimal("-15000"));
-  }
-
-  @Test
-  void
-      sell_respectsMinTradeFloorAndLeavesSubThresholdResidualToReserveRatherThanDumpingOnAtModelPosition() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00OVER", new BigDecimal("40000")),
-                    new PositionSnapshot("IE00ATTARGET", new BigDecimal("100000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00OVER", ZERO),
-                    new ModelWeight("IE00ATTARGET", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-44000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.trades())
-        .usingRecursiveComparison()
-        .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
-        .isEqualTo(
-            List.of(
-                new TradeCalculation(
-                    "IE00OVER", new BigDecimal("-40000"), new BigDecimal("0.000000"), OK),
-                new TradeCalculation("IE00ATTARGET", ZERO, new BigDecimal("0.100000"), OK)));
-  }
-
-  @Test
-  void sell_failsWhenTotalSellNeedExceedsSellableMarketValue() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("40000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("30000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-90000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    assertThatThrownBy(() -> engine.calculate(input, SELL))
-        .isInstanceOf(IllegalStateException.class)
-        .hasMessageContaining("Insufficient liquidity");
-  }
-
-  @Test
-  void buy_concentratesFreeCashIntoSingleAboveThresholdTradeInsteadOfSubThresholdSplits() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("315900")),
-                    new PositionSnapshot("IE00B", new BigDecimal("316100")),
-                    new PositionSnapshot("IE00C", new BigDecimal("368000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.32")),
-                    new ModelWeight("IE00B", new BigDecimal("0.32")),
-                    new ModelWeight("IE00C", new BigDecimal("0.36"))))
-            .grossPortfolioValue(new BigDecimal("1050000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("50000"))
-            .minTransactionThreshold(new BigDecimal("50000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    assertThat(result.trades())
-        .usingRecursiveComparison()
-        .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
-        .isEqualTo(
-            List.of(
-                new TradeCalculation(
-                    "IE00A", new BigDecimal("50000.00"), new BigDecimal("0.348476"), OK),
-                new TradeCalculation("IE00B", ZERO, new BigDecimal("0.301048"), OK),
-                new TradeCalculation("IE00C", ZERO, new BigDecimal("0.350476"), OK)));
-  }
-
-  @Test
-  void sellFast_weightsFastBucketSellsByOverweightBeforeTouchingSlowBucket() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00F1", new BigDecimal("250000")),
-                    new PositionSnapshot("IE00F2", new BigDecimal("210000")),
-                    new PositionSnapshot("IE00SLOW", new BigDecimal("240000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00F1", new BigDecimal("0.30")),
-                    new ModelWeight("IE00F2", new BigDecimal("0.30")),
-                    new ModelWeight("IE00SLOW", new BigDecimal("0.40"))))
-            .grossPortfolioValue(new BigDecimal("640000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-60000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of("IE00F1", "IE00F2"))
-            .build();
-
-    var result = engine.calculate(input, SELL_FAST);
-
-    assertThat(result.trades())
-        .usingRecursiveComparison()
-        .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
-        .isEqualTo(
-            List.of(
-                new TradeCalculation(
-                    "IE00F1", new BigDecimal("-45789.47"), new BigDecimal("0.319079"), OK),
-                new TradeCalculation(
-                    "IE00F2", new BigDecimal("-14210.53"), new BigDecimal("0.305921"), OK),
-                new TradeCalculation("IE00SLOW", ZERO, new BigDecimal("0.375000"), OK)));
-  }
-
-  @Test
-  void sellFast_liquidatesFastBucketFullyBeforeSpillingRemainderToSlowBucket() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00F1", new BigDecimal("250000")),
-                    new PositionSnapshot("IE00F2", new BigDecimal("210000")),
-                    new PositionSnapshot("IE00SLOW", new BigDecimal("240000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00F1", new BigDecimal("0.30")),
-                    new ModelWeight("IE00F2", new BigDecimal("0.30")),
-                    new ModelWeight("IE00SLOW", new BigDecimal("0.40"))))
-            .grossPortfolioValue(new BigDecimal("200000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-500000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of("IE00F1", "IE00F2"))
-            .build();
-
-    var result = engine.calculate(input, SELL_FAST);
-
-    assertThat(result.trades())
-        .usingRecursiveComparison()
-        .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
-        .isEqualTo(
-            List.of(
-                new TradeCalculation(
-                    "IE00F1", new BigDecimal("-250000"), new BigDecimal("0.000000"), OK),
-                new TradeCalculation(
-                    "IE00F2", new BigDecimal("-210000"), new BigDecimal("0.000000"), OK),
-                new TradeCalculation(
-                    "IE00SLOW", new BigDecimal("-40000"), new BigDecimal("1.000000"), OK)));
-  }
-
-  @Test
-  void sell_lastResortFullyLiquidatesTrappedSubThresholdOddLotsMostOverweightFirst() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00BIG1", new BigDecimal("300000")),
-                    new PositionSnapshot("IE00BIG2", new BigDecimal("300000")),
-                    new PositionSnapshot("IE00ODD1", new BigDecimal("40000")),
-                    new PositionSnapshot("IE00ODD2", new BigDecimal("40000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00BIG1", new BigDecimal("0.30")),
-                    new ModelWeight("IE00BIG2", new BigDecimal("0.30")),
-                    new ModelWeight("IE00ODD1", new BigDecimal("0.20")),
-                    new ModelWeight("IE00ODD2", new BigDecimal("0.20"))))
-            .grossPortfolioValue(new BigDecimal("680000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-660000"))
-            .minTransactionThreshold(new BigDecimal("50000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.trades())
-        .usingRecursiveComparison()
-        .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
-        .isEqualTo(
-            List.of(
-                new TradeCalculation(
-                    "IE00BIG1", new BigDecimal("-300000"), new BigDecimal("0.000000"), OK),
-                new TradeCalculation(
-                    "IE00BIG2", new BigDecimal("-300000"), new BigDecimal("0.000000"), OK),
-                new TradeCalculation(
-                    "IE00ODD1", new BigDecimal("-40000"), new BigDecimal("0.000000"), OK),
-                new TradeCalculation(
-                    "IE00ODD2", new BigDecimal("-40000"), new BigDecimal("0.000000"), OK)));
-  }
-
-  @Test
-  void sellFast_safetySpillFulfilsRedemptionTheBucketsStrandBelowMinTrade() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00FASTBIG", new BigDecimal("86000")),
-                    new PositionSnapshot("IE00FASTODD1", new BigDecimal("4000")),
-                    new PositionSnapshot("IE00FASTODD2", new BigDecimal("3000")),
-                    new PositionSnapshot("IE00SLOWODD", new BigDecimal("2000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00FASTBIG", new BigDecimal("3")),
-                    new ModelWeight("IE00FASTODD1", new BigDecimal("5")),
-                    new ModelWeight("IE00FASTODD2", new BigDecimal("4")),
-                    new ModelWeight("IE00SLOWODD", new BigDecimal("4"))))
-            .grossPortfolioValue(new BigDecimal("95000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-4816"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of("IE00FASTBIG", "IE00FASTODD1", "IE00FASTODD2"))
-            .build();
-
-    var result = engine.calculate(input, SELL_FAST);
-
-    assertThat(result.trades())
-        .usingRecursiveComparison()
-        .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
-        .isEqualTo(
-            List.of(
-                new TradeCalculation("IE00FASTBIG", ZERO, new BigDecimal("0.905263"), OK),
-                new TradeCalculation(
-                    "IE00FASTODD1", new BigDecimal("-4000"), new BigDecimal("0.000000"), OK),
-                new TradeCalculation("IE00FASTODD2", ZERO, new BigDecimal("0.031579"), OK),
-                new TradeCalculation(
-                    "IE00SLOWODD", new BigDecimal("-2000"), new BigDecimal("0.000000"), OK)));
-  }
-
-  @Test
-  void sell_withoutSoftLimitsSpreadsSellsAcrossOverweightFundsTowardModelRatherThanConcentrating() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("86151")),
-                    new PositionSnapshot("IE00B", new BigDecimal("262989")),
-                    new PositionSnapshot("IE00C", new BigDecimal("93572"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", ZERO),
-                    new ModelWeight("IE00B", new BigDecimal("0.50")),
-                    new ModelWeight("IE00C", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("442712"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-354775"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.trades())
-        .usingRecursiveComparison()
-        .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
-        .isEqualTo(
-            List.of(
-                new TradeCalculation(
-                    "IE00A", new BigDecimal("-86151"), new BigDecimal("0.000000"), OK),
-                new TradeCalculation(
-                    "IE00B", new BigDecimal("-201180.01"), new BigDecimal("0.139614"), OK),
-                new TradeCalculation(
-                    "IE00C", new BigDecimal("-67443.99"), new BigDecimal("0.059018"), OK)));
-
-    BigDecimal totalSold =
-        result.trades().stream().map(TradeCalculation::tradeAmount).reduce(ZERO, BigDecimal::add);
-    assertThat(totalSold).isEqualByComparingTo(new BigDecimal("-354775"));
-  }
-
-  @Test
-  void rebalance_whenModelNetDivergesFromFreeCash_warnsToCheckTheInputs() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("600000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("400000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1100000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("40000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    assertThat(result.warnings())
-        .extracting(CalculationWarning::type)
-        .containsExactly(REBALANCE_NET_CASH_MISMATCH);
-  }
-
-  @Test
-  void rebalance_whenModelNetMatchesFreeCash_raisesNoWarning() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("600000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("400000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1100000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("100000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    assertThat(result.warnings()).isEmpty();
-  }
-
-  @Test
-  void rebalance_whenTheGapIsExactlyTheMinTransactionThreshold_stillWarns() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("600000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("400000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1100000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("95000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    assertThat(result.warnings())
-        .extracting(CalculationWarning::type)
-        .containsExactly(REBALANCE_NET_CASH_MISMATCH);
-  }
-
-  @Test
-  void rebalance_whenAHardLimitBlocksPartOfTheBuySide_warnsTheNetWasNotAchieved() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("100000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("900000"))))
-            .modelWeights(
-                List.of(
-                    new ModelWeight("IE00A", new BigDecimal("0.50")),
-                    new ModelWeight("IE00B", new BigDecimal("0.50"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                    new PositionLimitSnapshot(new BigDecimal("0.20"), new BigDecimal("0.20"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, REBALANCE);
-
-    assertThat(result.warnings())
-        .extracting(CalculationWarning::type)
-        .containsExactly(REBALANCE_NET_NOT_ACHIEVED);
   }
 
   @Test
@@ -1917,24 +103,6 @@ class TradeCalculationEngineTest {
 
   // A rebalance needs something to rebalance against; with no positions there is nothing to warn
   // about either.
-  @Test
-  void rebalance_withoutAnyPositions_raisesNoCashWarnings() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of())
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(ZERO)
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    assertThat(engine.calculate(input, REBALANCE).warnings()).isEmpty();
-  }
 
   @Test
   void buy_withANegativeModelWeight_failsLoudly() {
@@ -1954,171 +122,5 @@ class TradeCalculationEngineTest {
 
     assertThatThrownBy(() -> engine.calculate(input, BUY))
         .isInstanceOf(IllegalArgumentException.class);
-  }
-
-  @Test
-  void buy_withNoPositions_recordsNoPositionsReason() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of())
-            .modelWeights(List.of())
-            .grossPortfolioValue(ZERO)
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(ZERO)
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    assertThat(result.noTradeReason()).isEqualTo("No trades: mode=BUY, reason=noPositions");
-  }
-
-  @Test
-  void buy_withFreeCashExactlyAtThresholdButNoHeadroom_recordsNoPositionUnderTargetReason() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("400000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(new BigDecimal("500000"))
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("5000"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                    new PositionLimitSnapshot(new BigDecimal("0.40"), new BigDecimal("0.4001"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, BUY);
-
-    assertThat(result.noTradeReason())
-        .isEqualTo("No trades: mode=BUY, reason=noPositionUnderTargetBeyondThreshold");
-  }
-
-  @Test
-  void sell_withFreeCashExactlyAtNegativePointZeroOne_recordsNoCashShortfallReason() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("10000"))))
-            .modelWeights(List.of(new ModelWeight("IE00A", new BigDecimal("1.00"))))
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-0.01"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.noTradeReason())
-        .isEqualTo("No trades: mode=SELL, reason=noCashShortfall, freeCash=-0.01");
-  }
-
-  @Test
-  void sell_withShortfallOneCentBeyondSellableMarketValue_stillFullyLiquidatesWithoutThrowing() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("40000"))))
-            .modelWeights(List.of())
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-40000.01"))
-            .minTransactionThreshold(new BigDecimal("5000"))
-            .positionLimits(
-                Map.of(
-                    "IE00A",
-                    new PositionLimitSnapshot(new BigDecimal("0.99"), new BigDecimal("0.999"))))
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.trades().getFirst().tradeAmount())
-        .isEqualByComparingTo(new BigDecimal("-40000.00"));
-  }
-
-  @Test
-  void sell_withCapExactlyAtThresholdTolerance_stillSellsInsteadOfBeingExcluded() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("99.99"))))
-            .modelWeights(List.of())
-            .grossPortfolioValue(new BigDecimal("1000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-200"))
-            .minTransactionThreshold(new BigDecimal("100"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.trades().getFirst().tradeAmount())
-        .isEqualByComparingTo(new BigDecimal("-99.99"));
-  }
-
-  @Test
-  void sell_withATiedSubThresholdSplit_eliminatesTheFirstTiedPositionNotTheLast() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(
-                List.of(
-                    new PositionSnapshot("IE00A", new BigDecimal("1000000")),
-                    new PositionSnapshot("IE00B", new BigDecimal("1000000"))))
-            .modelWeights(List.of())
-            .grossPortfolioValue(new BigDecimal("2000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-1500"))
-            .minTransactionThreshold(new BigDecimal("1000"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    var tradeA =
-        result.trades().stream().filter(t -> t.isin().equals("IE00A")).findFirst().orElseThrow();
-    var tradeB =
-        result.trades().stream().filter(t -> t.isin().equals("IE00B")).findFirst().orElseThrow();
-    assertThat(tradeA.tradeAmount()).isEqualByComparingTo(ZERO);
-    assertThat(tradeB.tradeAmount()).isEqualByComparingTo(new BigDecimal("-1500.00"));
-  }
-
-  @Test
-  void sell_withAmountExactlyAtThresholdTolerance_stillAllocatesInsteadOfDroppingTheRunner() {
-    var input =
-        FundTransactionInput.builder()
-            .fund(TUV100)
-            .positions(List.of(new PositionSnapshot("IE00A", new BigDecimal("1000000"))))
-            .modelWeights(List.of())
-            .grossPortfolioValue(new BigDecimal("2000000"))
-            .cashBuffer(ZERO)
-            .liabilities(ZERO)
-            .freeCash(new BigDecimal("-99.99"))
-            .minTransactionThreshold(new BigDecimal("100"))
-            .positionLimits(Map.of())
-            .fastSellIsins(Set.of())
-            .build();
-
-    var result = engine.calculate(input, SELL);
-
-    assertThat(result.trades().getFirst().tradeAmount())
-        .isEqualByComparingTo(new BigDecimal("-99.99"));
   }
 }
